Virtual method calls are basically working.  This required a few different pieces of functionality:
1) In Elsa, we now maintain a map from a class to its subclasses.

2) Also in Elsa, when a method is typechecked in
cc_tcheck.cc:Declarator::mid_tcheck() we search up the inheritance
hierarchy for superclasses containing a method that we override and
flag the current method virtual if the superclass method is virtual.
This works because we typecheck the hierarchy in descending order.  If
there is some funniness with ambiguous inheritance, I don't deal with
it yet.

3) In cc_qual, when inserting inference edges for a function call in
qual_walk.cc:E_funCall::iqual(), if the static-typed method is
virtual, we also search down the hierarchy from the class that the is
in, looking for all other methods that inherit from it.  The virtual
method edge insertion will still fail if we have a class in another
translation unit that inherits that we don't know about.  This is
because we typecheck files separately in oink_main.cc.  We need a new
mechanism to fix that.