Added typechecking for E_addrOf so that it makes pointers to members when it should.  Before, was just making normal pointers.  All of the testing is in cc_qual.
Added a utility function belowE_grouping to the Expression AST node.
It removes layers of parens (E_grouping expressions) until it gets to
something that is not an E_grouping.

Added a makePtrToMemberType() factory method parallel to makePtrType()
on both the typechecking Env class and type factory classes.

In computeArraySizeFromCompoundLiteral() fixed places where we now
need an extra "->expr" due to the ICExpression being wrapped around
the Expression nodes.  This is only on when GNU_EXTENSION is set, so
Scott didn't catch it.

In E_binary::itcheck() added an assertion that we aren't making a
pointer to member to a reference except where it is explicitly
allowed.  Replaced references to the mysterious BIN_ARROW_DOT operator
with BIN_DOT_STAR.