Changes to cc_tcheck.cc E_addrOf::itcheck():
If there is a paren such as: &(...) then the internals are checked as
if this->..., and only making a normal pointer is allowed.  Therefore
fallthrough instead of giving an error.

Put the call to makePTMType at the top once instead of mixing it in
with the normal pointer code that is deciding between functions and
data pointers.

Replace
  CompoundType *inClass0 = dynamic_cast<CompoundType*>(var0->scope);
with
  inClass0 = var0->scope->curCompound;