Merged in all the differences from overload1start to merge_to_overload_branch1 in the main branch (trunk).
In areYouOrHaveYouEverBeenATemplate() we also now return true if the
type is a dependent type.

Operator overloading is still off (was on in the main branch when
merged in), as it fails for the [] operator since it is being
converted to a composition of "+" and "*" and then the "+" is checked
for overloading and now candidates are found.  Consequently had to
turn off the operator overloading test in/t0127.cc .

Now pass -tr doOverload on most of the tests in regrtest.