Do ctor overloading:
    Now add implicit ctor and implicit copy ctor for classes that don't
    have them.

    Added places in cc_tcheck.ast to put the ctor overload resolution
    vars.

Renamed Env::makeDestructorFunctionType() to
Env::makeCDtorFunctionType(), since it does both now.

Created and implemented the semantics for a flag type enum
TypeToStringFlags for indicating if toString() and toCString() on
types should be 1) as before, TTS_NONE, or 2) canonical, TTS_CANON;
the canonical version is suitable as a poor-man's name mangling for
linker imitation.

(With suggestion from Scott) fixed bug that addVariable() was doing
some things registerVariable() should be doing.

Made BIN_COMMA allow propagation of a reference type up from the right
side so that "&(3, a)" works.