Now make ctors and dtors for stack allocated, function scope variables.  Big tests don't pass again.
All of the typechecking of Declarator::init was moved into
elaborateCDtors() to make this work.  This is natural, as the init is
part of dealing with the ctor.

Also to do this, had to track the difference between typechecking
Declarators that are parameters and those that are not, as parameters
take a copy ctor.  Parameters are still not elaborated with ctors or
dtors.

Shadow typedef vars added for typedef vars that get hidden by normal
vars with the same name.

In Env::instantiateClassTemplate() the instName for an instantiated
templatized class no longer has the names of the template arguments in
angle brackets folded into the string of the name for the typedefVar
for the class.  This name is used to make a fully qualified PQName
name for this class (where the template arguments are kept separate)
and is therefore no longer considered to be for "debugging purposes
only".