Added makeCopyCtorBody() and support functions for making the bodies of the implicit copy ctors; in addCompilerSuppliedDecls() this is now called if env.doElaboration is on, otherwise if env.lang.hasImplicitStuff is on, only a declaration is made; addCompilerSuppliedDecls() is not even called at all if not env.lang.hasImplicitStuff.
Added makeCopyAssignBody() and support functions for making the bodies
of the implicit copy assign operator; this is not yet turned on nor
tested but compiles.

In the ctor for Env, doElaboration is turned on if
lang.hasImplicitStuff.

Env::make_PQ_fullyQualifiedName() is broken out into smaller, more
reusable pieces.

Added assertion to makeCtorExpr(), makeCtorStatement(), and
makeDtorStatement() that they are only called under env.doElaboration;
this way we can test they are not called from C when AST nodes that
would trigger them in C++ are encountered; added running an additional
C test in/c/t0005.c to use all of these AST nodes in C mode.

All of the calls to the above-mentioned functions are now guarded by
env.doElaborate tests.

Added assertion to class Expression that when cloned should have had
all of its ambiguities resolved.

Fixed bug where was testing for an existing copy assign operator by
counting the number of arguments; should be testing for 2 not 1.

Commented out pretty printing test t0034.cc since template
instantiations don't seem to be pretty printing right.