Removed silly redundancy where makeNoArgCtorBody() was making no-arg MemberInitializers for superclasses and members when this is completely unnecessary since when the Function is typechecked completeNoArgMemberInits() is going to do it.
In completeNoArgMemberInits() when we are about to make a no-arg ctor
for a member, we first check that it is of CompoundType.  If not, we
don't.  This seems to be the intention of cppstd 12.6.2 para 4.

Fixed silly bug where I was ignoring the return value of
FakeList<>::prepend instead of resetting the list to be that return
value.

In Function::tcheck() now check if the function is a ctor that needs
its inits tchecked by looking at its name instead of checking if it
has non-NULL inits, as I need to now typecheck the MemberInitializers
even if there are none (to add the no-arg ones that are missing).

Removed commented out old first try at making a no-arg ctor body.

No good testing for these implicit function bodies yet, otherwise the
above bugs would have been fixed the first time.