Dtor elaboration is turned on!: both adding the dtor body when missing and adding the superclass and member dtors at the end of the dtor body, whether user supplied or implicitly defined.
Had to turn off some pretty printing test since when you put the
superclass/member dtor calls at the end of a dtor, they get pretty
printed, and then when you parse it and do it again, you get two
copies.

Also had to turn off one of the tests from the standard,
in/std/7.3.3j.cc, since the elaborated dtor ends up in the
collectLookupResults resulting in a pseudo-failure.