Env::Env() now turns on doOverload and doOperatorOverload only if lang.allowOverloading is also on.
The self-name is only added to a CompoundType if
lang.compoundSelfName, which is on for C++ and not for C.

Scope::fullyQualifiedName() now prints template arguments as part of
the name of classes that are in the qualification string as we go up
the nesting.

We now allow cast to pointers as well as ints as part of constant
expressions.