Added bool Function::implicitlyDefined flag that is set for those functions with bodies of completely artificially generated AST.
Fixed bug in isCopyAssignOp() where I was checking against the wrong
parameter to see if it was a suitable parameter for a copy assign
operator.  Previously I had had a bug of counting the number of
parameters wrong, which had been fixed but I didn't think to check
later in the code when I got the parameter out to examine that I was
getting the wrong one.  Also fixed another bug in isCopyAssignOp()
where I had the sense reversed of the test of whether the parameter in
question, if atomic, was equal to the CompoundType in question.  Also
simplified the end of the function.