The AST now only has c/dtorStatements instead of Exprs, so changed the S_return::ctorExpr to a ctorStatement.
Added elaboration for Handler: a globalVar to match the global in the
E_throw (some other analysis must hook them together, but the
data-structures should be parallel); it gets a dtorStatement just as
an E_throw's global gets a ctorStatement.  We consider a Handler to be
a call to an anonymous inline immediate function, so we have a
localArgument expression which is the local temporary in the case of
catch by value; this also entails that a Handler get a
FullExpressionAnnot, which it now does; also the handler declarator is
now typechecked as if it were a parameter.  Added a
Env::makeCatchClauseVarName() to generate artificial names for Handler
global variables.

Added a globalCtorStatement for E_new that ctors the global exception.

Now elaborate a dtorStatement for E_delete.

****

At times when CompoundTypes are passed by value, one must elaborate
temporaries and copy ctor calls targeting them.  Calls to the copy
ctor require an argument which is at times an existing expression
(such as the expression of an S_return).  In order to prevent the AST
from becoming not a tree by the necessary placement of this expression
into the argument of the copy ctor, I SET TO NULL THE ORIGINAL; that
is, elaboration is mutating the AST.

The other solution might be to instead clone the expression, but this
produces bad results if it contains a FullExpressionAnnot and has
already been typechecked.  Therefore I ADDED CODE TO DO ALL OF THIS
CLONING but IT IS LEFT COMMENTED OUT and not used; comments about
cloning are left in place as if it is still happening; we should
decide on this, but what I do now seems to work.  Added a
cloneFakeList_ArgExpression(FakeList<ArgExpression> *args0) but don't
use it because cloning is causing problems when interacting with
FullExpressionAnnot.

As part of this no-cloning scheme, moved elaboration of S_return back
up to the top before typechecking.

****

Added serialNumbers to types to aid in debugging, conditioned on the
compile time flag USE_TYPE_SERIAL_NUMBERS and the environment variable
PRINT_TYPE_SERIAL_NUMBER; the configure flag to turn on the compile
time option is -useTypeSerialNumbers.  We also made some modifications
to the way printTypedAST prints variables and types so they have more
information.  This causes a bug in that the comments no longer nest
correctly when types are pretty printed.

Had to comment out three pretty printing idempotency tests since my
modifications of the ast during elaboration seem to cause whitespace
changes.

****

elaborateCallByValue() factored out for use 1) in function/ctor call
parameter passing by value and 2) for a Handler where the exception is
caught by value.

Factored makeDtorExpr() out from makeDtorStatement(); now make the
dtor expression as a method call instead of a function call, since
dtors have a "this" argument.

Rearranged elaborateCDtors(Env &env) as the logic for dealing with the
initializers was terrible.

****

Fixed bug in makeCtorExpr() where variable named "this" was handled
incorrectly.

Fixed bug in elaborateCallSite() where artificial ctors where having
temporaries made for them, even though they already get a target,
which is the point of their existence; this entailed adding a bool
artificial parameter to the signature.

Fixed bug in elaborateFunctionStart() where the <retVal> variable was
not of type _reference_ to the ret type of the function.  This
required other changes throughout.

Fixed bug in E_new::elaborate() where we were not elaborating a global
variable for the exception if it was not of compound type.

Fixed bug in E_throw::elaborate() where wasn't cloning the type of the
expr before re-using it to make the globalVar.

Fixed bug that Function::dtorStatement was not being typechecked.

Seems I fixed a bug where E_funCall::retObj was not being typechecked
after it was elaborated into existence.

With Scott, fixed a bug where the return type "Type *rt" of E_deref
was not being cloned.  Also elaboration now only done in the case of
the argument not causing an error by being the wrong type.

****

Scott re-factored MemberInit::tcheck() in the previous check-in making
it much cleaner but I had to change it again to pull
tcheckArgExprList(args, env); back out of the top so that in the cases
where the arguments end up in a ctorStatement they are not typechecked
twice.  Despite the fact that typechecking is supposed to be
idempotent, it is best not to test fate, as I'm having lots of hard to
track down bugs.

Scott changed fields to use the new feature that you can put an init
expr in for class fields in .ast files now.  Scott also added use of
the new "public(field,owner)" feature.