It used to be that this would result:
cich:~/qual_base/elsa$ ./ccparse cndecl1.cc
cndecl1.cc:3:7: WARNING: there is no action to merge nonterm Condition
Aborted

when run on this input:

struct A{};
int main() {
  if (A *a = 0) {
  }
}

This is fixed by adding ambiguity resolution mechanism to the
Condition grammar nonterminal.  The above example is checked in as
in/d0052.cc and added to regrtest.