In oink, the type integrity checker/visitor is now on, which is checking that there is at most one path to any type for those that are used in a dataflow analysis.  This found many instances where cloneType-s needed to be introduced, and they were.  One place was in Expression::mid_tcheck() which clones the return type of the itcheck_x() before annotating the Expression node with it.  This removes the necessity for many cloneType-s that were previously in some of the itcheck_x() methods, and so they were removed.
There is now a Variable::toString() that prints the name and the type
of the variable; therefore we no longer need the Type::toString(char
const *name) method; this makes uniform toCString() and toMLString().

Changed USE_TYPE_SERIAL_NUMBERS and typeSerialNumbers to
USE_SERIAL_NUMBERS and serialNumbers respectively since Variables and
Types now both have them.

Added guards to check that if USE_SERIAL_NUMBERS is defined, it is 0
or 1.

Scott added that a Declarator now prints out its context when the AST
is dumped.

ElabVisitor::makeFunction() now sets the new Function's receiver to
that of its funcType rather than that of the input var's function
type's.

E_new::ctorVar is now printed by custom debugPrint code; I think this
was to tolerate situations in which it can be NULL.

Scott modified regrtest so that the basic testparse tests are in a
block and delimited by special comments so that I can get them out
with perl and automatically keep my tests in oink up to date.