Implemented class template partial specialization for int and type template-parameters; still not done are template template-parameters, non-int object parameters (pointer, reference, enum), and separate function definitions for functions declared within templatized classes.
Since class template instantiations must be unambiguous we manage to
re-use as-is the tournament code for function overload resolution.
The new class TemplCompoundType holds the templatized CompoundType
candidates and thereby satisfies the API for the tournament.

TemplateInfo, STemplateArgument, and the constructed Type hierarchy
all have cooperating new atLeastAsSpecificAs() methods that do the
"matching" or "type unification" between a list of template arguments
and a specialization pattern (template arguments with variables) of
partial specializations.  These methods serve to both check that
unification is possible and to compute the bindings necessary to
effect it.  Note the asymmetry: one input is a list of actual
object/types and the other of a list of patterns: objects/types with
variables that can be bound during the unification process.  Note that
bindings are built as required and those previously built are checked
for consistency; inconsistent bindings are therefore not possible.

Classified all the combinations of states that TemplateInfo objects
can be in as:
        1) primary,
        2) partial spec,
        3) complete spec or instantiation, and
        4) mutant state for instantiations created with
           type-vars as arguments.

These states are written up extensively in the comments for class
TemplateInfo along with an extensive exposition on the operation of
templates in general and the relationship between arguments and
parameters.

Distinguishing TemplateInfo states from one another can be quite trick
and is still not quite done correctly.  The method
argumentsContainVariables() helps with this but needs to be modified.
The method argumentsAllVariables() is probably misguided and should be
removed, but we check it in for now.

The method instantiateClassTemplate() had in-lined code for adding
variables to the template scope during the instantiation of a primary;
this was factored out into insertBindingsForPrimary() and a similar
function for partial specializations was created,
insertBindingsForPartialSpec().

Scope::fullyQualifiedName() had to be extended to produce reasonable
mangled names for object template-parameters.

Added some gdb() methods for TempateInfo and STemplateArgument.

---

Fixed a bug that
  CANDIDATE *tournament(RESOLVER &resolver, int low, int high, CANDIDATE *dummy)
would go into an infinite loop if low was greater than high, which
would happen if the client selectBestCandidate() function were called
on an empty set

Fixed a bug that selectBestCandidate() would call the tournament on an
empty set instead of just returning NULL.

---

Removed old aborted attempt at implementation by Scott.