Function template instantiation, including complete specializations and partial specializations, is implemented.  One major change is that instantiateClassTemplate() was generalized to handle both class and function templates and renamed to instantiateTemplate().
Forwarding for function templates is simply not handled.

Implied function template arguments from function arguments is not
implemented.

Selecting between the various overloaded function template primaries
for a given function template specialization to be filed under is done
using FunctionType::atLeastAsSpecificAs() and giving an error if more
than one satisfies it.  I don't know if this the real rule, as the
spec is not very clear on this situation.

Actually instantiating overloaded function templates is not
implemented; we continue to duck it as before.

****

Moved the template-ness, that is the "TemplateInfo *" member, of a
CompoundType to Variable so that both Functions and Compounds can be
templatized.  This data member of Variable is now called templInfo and
can accessed through the templateInfo() and setTemplateInfo() methods,
which are paralleled on CompoundType so that code which is used to
accessing the template info from a CompoundType can still do so; the
implementation goes through the typedefVar member to get at the real
TemplateInfo object.

In six places, all marked "FUNC TEMPLATE LOSS" in a comment, the
functionality of the code had to be weakened in some way due to the
fact that FunctionType no longer has a templateInfo member to query.
None of these areas looked critical.  The client code in question that
no longer works is checked in commented out, to be re-implemented some
new way perhaps.

One of them is that template parameters no longer pretty print, and so
the relevant pretty printing in/t0026.cc test had to be turned off.

Moved the line below in Env::makeNewCompound() to below the part where
the typedefVar is created, as the templateInfo now lives there in
reality, although the CompoundType API makes it look like it is still
in CompoundType.
  ct->setTemplateInfo(takeCTemplateInfo(name));


Two changes in D_func::tcheck():

1 - moved
  ft->templateInfo = templateInfo;
to
  dt.var->setTemplateInfo(templateInfo);
at the bottom of the file because the variable to which the template
info is attached does not exist until after
  if (templateInfo &&
      (templateInfo->isCompleteSpecOrInstantiation() ||
       templateInfo->isPartialSpec())
      ) {
    dt.dflags |= DF_TEMPL_SPEC;
  }
  base->tcheck(env, dt, inGrouping);

2 - Above one can also see the new declaration flag DF_TEMPL_SPEC
being put onto template specializations.

In Env::createDeclaration() we now prevent variables with this flag
from being added to an overload set; this would have put it in the
same overload set as its primary when it should instead go into the
instantiations list of its primary.
  if (!(dflags & DF_TEMPL_SPEC)) {
    overloadSet->addMember(newVar);
    newVar->overload = overloadSet;
  }


Two changes to Env::lookupPQVariable():

1 - There is a new LookupFlags LF_TEMPL_PRIMARY.  Under this flag the
lookup will return a template primary rather than instantiating it.

2 - The old method
  Env::lookupPQVariable(PQName const *name, LookupFlags flags)
is now just a wrapper around a call to the real method
  Env::lookupPQVariable(PQName const *name, LookupFlags flags, FunctionType *signature)
with signature set to NULL.  The additional FunctionType parameter is
for disambiguating between various function template primaries when
they are overloaded and LF_TEMPL_PRIMARY primary is in effect.  This
happens when a function template primary is looked up so that a
function template specialization can be filed under it.


Renamed class TemplCompoundType to class TemplCandidates, as now
CompoundType-s or FunctionType-s can be candidates.  The candidates
are now Variables.

Added a method bool Scope::isGlobalTemplateScope() which is supposed
to test if the scope is a template scope within the global scope.  In
Scope::registerVariable() we now set the flag DF_GLOBAL if we are in a
global scope (as before) or if we are in a global template scope.
This is critical to the correct operation of the linker imitator in
Oink.

Generalized Scope::fullyQualifiedName() to work for object template
parameters ("non-type non-template template parameters") that is, say,
ints.

In TD_func::itcheck() we look to see if the function template
declaration is in fact a declaration for a specialization and if it is
we add it to the instantiations list of the primary.  This parallel
the same functionality for class templates in TS_classSpec::itcheck(),
though I wonder why that wasn't done in TD_class::itcheck().

Added method CompoundType::getTypedefVar() that checks that the
following invariant holds every time the typedefVar is accessed.
  CompoundType::typedefVar->type->asCompoundType() == this
This invariant is critical to the correctness of the
CompoundType::set/templateInfo() methods, as the TemplateInfo object
is now on the typedefVar, not the CompoundType.
CompoundType::set/templateInfo() uses getTypedefVar() and therefore
indirectly checks this invariant every time the TemplateInfo object is
accessed.

TemplateInfo::gdb() now works for both class and function templates.

Fixed bug in PointerType::atLeastAsSpecificAs() where references and
pointers were being considered as if they are interchangeable.

TypeFactory::makeSimilarFunctionType() had to be changed to remove the
copying of the template info when a similar function type is made.
This results in a change in the client code where
makeSimilarFunctionType() was being used:
  void Env::makeUsingAliasFor(SourceLoc loc, Variable *origVar)
  {
    FunctionType *newFt = tfac.makeSimilarFunctionType(SL_UNKNOWN, oldFt->retType, oldFt);
    . . .
    // additional code below added to make up for its removal
    // from makeSimilarFunctionType
    if (origVar->templateInfo()) {
      newVar->setTemplateInfo(new TemplateInfo(*origVar->templateInfo()));
    }

Added method
  OverloadSet::findTemplPrimaryForSignature(FunctionType *signature)
that given a function signature will hunt through the template
primaries in the OverloadSet and find the one primary template that
the signature is atLeastAsSpecificAs(); returns NULL if there is none
and calls xfailure() if there is more than one.  This is used to find
the primary in an overload set when a function template specialization
is tchecked.  I don't know if this fits the behavior of the spec, as
the spec is not very spec-ific.  Note that the xfailure() should
really be a user error of some kind, but for now I want it to fail
loudly.

Changed the forward function template declaration in/d0005.c to a
definition, as we don't deal with forward function templates yet.

Fixed a bug in in/t0026.cc: non-compound `int' doesn't have fields to
access:
  template <class T>
  int f(T &t) {
    t.x;  // this line commented out since it is wrong given the instantiation
  }
  . . .
  f<int>(8);