This check-in is to get Oink in sync with the changes that Scott just made that split PointerType and ReferenceType.
Renamed makeRefType() to makeReferenceTypeIdem() because:
        1) it sounded too much like makeReferenceType() which means
something else; it is not good to have two names with different
semantics that differ only in choice of abbreviation
        2) the point of this function is that it is idempotent where
as makeReferenceType() is not, so I made that obvious in the names

Added a call to cloneType() in E_funCall::inner2_itcheck() that Scott
had omitted.

In Declaration::tcheck() added a check that if the first declarator in
the decllist produces a NULL dt1.var upon being typechecked, that we
immediately return.  In oink/test/ptr-to-memberI.cc I check what
happens if you try to make a pointer to a reference member and if I
don't have this check I get a segfault instead of the user error
message that is pending.

Extended BaseType::getAtType() to handle Array, PointerToMember, and
Function (where it is idempotent); Oink had an identical function and
I didn't want duplicated functions with very similar semantics, but
Oink required these extension, so I made them.