Fixed function template instantiation to work correctly; in particular, this makes (self-)recursive function templates work.
Implemented function template forward declaration which also makes
mutually recursive function templates work.

Major rewriting of matchtype.cc so that 1) it caches the type pairs
that it has considered, preventing infinite loops, and 2) has the
nomenclature cleaned up and is much better factored.  This solution is
still wrong as is evidenced by the following example.

    template<class T> T f(T x, T y) {
      return f(y, 0);
    }

Matching the argument with the first argument/parameter pair you get
the binding T->T; matching the second we attempt to match an int to a
T and since T is already bound we will recurse into the same state on
its bound value, which is itself, notice the infinite loop and then
stop without having bound T->int, the most specific result.  Maybe we
do need union find for this.

---

Declarator::Tcheck has a new member and APIs for Function::tcheck(),
Declaration::tcheck(), and Env::createDeclaration() has a new argument
  bool reallyAddVariable
which says if it should we really add the variable to the scope after
it creates it.  This is false for typechecking both the declaration
and definition of function templates, as the variable should live in
the instantiations list of the primary, not in any particular scope.

Declarator::Tcheck has a new member APIs for Function::tcheck() and
Declaration::tcheck() have a new argument
  Variable *prior
which means use this prior definition rather than creating another
variable during typechecking; this is checked for at the very top of
D_name_tcheck() rather attempting to go through the body and that of
createDeclaration() without mishap (which turned out to be quite
difficult) when we already know what we want.  For now it is only
non-NULL only in the case of typechecking the body of a function
template definition; when it is non-NULL, reallyAddVariable should
also be false.

To Env::error() removed 'bool disambiguates' and replaced it with a
flag
  ErrorFlags eflags
which subsumes the functionality of the previous.  Added a new kind of
ErrorFlags EF_STRONG which means it may not be suppressed but it is
not disambiguating.

In matchtype.cc fixed a bug (by refactoring the code away) where we
weren't considering the possibility of type variables in the left
argument of the inClass member in the case of matching
PointerToMember-s.

Removed all references to [Base]Type::equals() in matchtype.cc;
matchtype is effectively doing something that subsumes or is a
generalization of the notion of equality, so to drop into
Type::equals() part way through is a real incongruity; note that we
still use AtomicType::equals().

Added to class Env but left commented out a new member
  SObjStack<TemplateDeclaration> templateDeclarationStack
which is the stack of TemplateDeclarations we are inside of (which
include function template definitions); pushed at the top of
TemplateDeclaration::tcheck() and popped at the bottom.  The
motivation is that inside the definition of a template body, you are
not in "real" code, you are in "might-be" code, a sort of never-never
land for typechecking; you might want to change your behavior in such
situations.  This is all not used for now but I leave it checked in
and commented out.

Added a tracing flag 'topform' which counts the top level forms and
prints them out as you enter them in TranslationUnit::tcheck().
Turning on this tracing flag and setting a breakpoint right after the
print is a handy way to start typechecking at the top of a
particularly interesting top level form.

In regrtest changed in/t0192.cc from failparse to testparse as it
works now.  Added d0056.cc and d0057.cc to test recursive function
templates and added d0058.cc to attempt to test type inference all in
one place, though it probably contains nothing that does not duplicate
other tests.