There are now three distinct typechecking modes: TTM_1NORMAL for normal typechecking, TTM_2TEMPL_FUNC_DECL for checking the parameters of function template definitions where we instantiate templates but only their primaries and the instantiation may have type vars in the arguments, and TTM_3TEMPL_DEF for checking template body definitions where an instantiation request just returns the primary. Env::getTemplTcheckMode() consults two different stacks of various AST nodes and figures out which mode you are in; see its implementation for details.
Another overhaul of matchtype.cc.  There are now three matching modes:
MM_BIND: generate (and verify) bindings, MM_WILD: vars on the right
are wildcards that match anything but don't generate bindings, and
MM_ISO: types are verified to have isomorphic sets of variables; vars
on the right generate (and verify) bindings to vars on the left, but
nothing else.  Mostly this is tested well for TypeVariables and not
well for int variables; other kinds are not even implemented.

In D_func::tcheck() fixed bug that we were stomping on the dt.var's
template info with the new one even if it already had a perfectly good
one from a previous declaration and we were re-using the var; now if
we re-use the var, we re-use the template info.

Functions printBindings() and bindingsGdb() are consolidated into
bindingsGdb() and updated for new bindings class.