Added a big assertion to Declaration::tcheck() to check the semantics of the way function template primaries and non-primaries are being added to the namespace after the typechecking of the Declarator.  This found a lot of bugs.
---

Fixed the bug that AtomicType::equals() was doing the wrong thing for
atomic types that were typechecked in mode TemplTcheckMode
TTM_2TEMPL_FUNC_DECL (the parameter list of a function template): they
might not be '==', but still be isomorphic if they were instantiated
with a template argument that is the template parameter (type
variable) of the function template.

This required using type matching and so I had to make a global
TypeFactory so that we could get at it from within equality testing of
CVAtomicType().

---

Fixed a really bad bug in match type in MM_ISO mode where it was
basically completely wrong.  When it found a typevar on the right that
was already bound, it was testing that the type it was bound to was
the same as the type on the left.  First, this is likely to be wrong
under schemes that allow type cloning.  Second, this is definitely
wrong when we are actually replacing the left type in the same
function with another type that has modified cv-qualifiers on it,
which does a shallow clone.  The solution was to dig down and get the
typevars and test them for equality.

---

Created and switched to using TemplateInfo::isNotPrimary() instead of
!isPrimary(), as isNotPrimary() doesn't assertion-fail on a mutant
TemplateInfo.

Made TemplateInfo::instantiations private so that I could find all of
the places where we were using instantiations.append() and replace
them with addInstantiation() which sets the myPrimary correctly of the
thing being added to the instantiations list, as well as checking a
lot of things in setMyPrimary(), which also now checks even more
things.  This was to fix in the most general way the bug that complete
specializations were not having their myPrimary field set.  This also
required that in two places a templatized variable be added to the
instantiations list after its arguments had been filled in, instead of
before, since only after the arguments are there does it look like it
is not a primary and therefore ok to put into an instantiations list.

TemplateInfo::instantiations is now private so switched to using
getInstantiations() to get it when needing to iterate over it.

Commented out TemplateInfo::getMyPrimary() and replaced with
TemplateInfo::getMyPrimaryIdem() which is idempotent on primaries and
never returns NULL.

Switching to getMyPrimaryIdem() fixed a bug in
MatchTypes::match_TInfo() where we were doing something strange in the
case of a being a primary where we were saying it didn't match b, even
though it seems b could have been the same as a.

---

Fixed a bug in MatchTypes::match_STA() where we were incomplete when
matching STA_REFERENCE-s against each other.  STA_REFERENCE is being
maintained on an as-needed basis.  mode == MM_WILD is now right I
think, but MM_ISO is implemented just like MM_WILD instead of creating
a binding and checking it, and MM_BIND is not implemented at all, so
both of those last two are wrong.

---

Variable::getOverloadSet() checks that the type of 'this' is a
FunctionType; this caught a bug.  OverloadSet::addMember(Variable *v)
now asserts that its argument is a FunctionType.

Made another version of OverloadSet::findByType that infers the
receiverCV from the first argument, as this is the usual case.
  Variable *findByType(FunctionType const *ft, CVFlags receiverCV); // original
  Variable *findByType(FunctionType const *ft); // new

---

Added ERROR(8) case to in/d0053.cc that tests that two different
classes that have the same template signature fail to match.  This was
working but I wasn't convinced that it was.