Instead of ignoring forwarded function template specializations, we now enter them into the instantiations list of the primary and hook their definitions into them later.  This is a step towards fully correct implementation; all tests pass so I'm checking in.
---
In in D_name_tcheck() if the name of the Declarator was a PQ_template,
we now save its template arguments in the new
Declarator::Tcheck::ASTTemplArgs so that later in D_func::tcheck()
(which is suspended on the stack at the time) when the template info
is created for the variable we can convert them into
STemplateArguments and attach them.

We put forward declarations of function templates into the
instantiations list of the primary if no MM_ISO matching declaration
is already there.  In TD_func::itcheck() if we are typechecking the
definition of a function template specialization that has already been
declared we point the declaration's funcDefn to the definition
Function.