We now go back and re-instantiate forwarded function templates that were only declared at instantiation time.  This is mostly initiated from TD_func::itcheck() when the function template definition is encountered and is semantically similar to instantiateForwardClasses(), though differs in important details, particularly in the flow through instantiateTemplate(). Env::provideDefForFuncTemplDecl() and Env::instantiateForwardFunctions() were created to parallel instantiateForwardClasses() for this purpose.
---

Changed the MatchTypes::MFlags to an int depth for all of the APIs.
This is because the only use it had was to tell if we were at the top
of a type or not.  This was used for two purposes: 1) do const and
volatile matter for type matching?, and 2) do pointers and arrays
match each other?  It turns out that the questions do not have the
same notion of "top" in the presence of a top-level ref type.  Below a
top level ref is still top for purposes of the second question, but
not for purposes of the first.  Thus, we now have 3 levels, which,
being linearly ordered, are better represented as an int rather than
two flags.  This eliminates the possibility of easily adding other
flags, but I it is much better to solve the problem you have than
problems that may never occur and can be fixed later anyway.

The dummy singleton variable instFwdClassesTDSDummy is renamed to
mode1Dummy and moved into Env, since it is now used in more than one
place to temporarily convert typechecking mode TTM_3TEMPL_DEF into a
TTM_1NORMAL for some sub-tree of the typechecking process.  The old
use was to return to mode TTM_1NORMAL for purposes of instantiating
forward class templates while still under the template class
definition.  The new uses are 1) instantiating forwarded function
templates, which is very similar to the above, and 2) typechecking
complete specializations of function templates, which are real code,
not template code.  The new name of this dummy singleton reflects the
generality of the its new uses.

Generalized Env::lookupPQVariable_primary_resolve() [and
findTemplPrimaryForSignature() which it uses] to take a
MatchTypes::MatchMode matchMode argument.  It is now used in 3 places
with slightly different semantics which require different mode
matching, though I'm still not sure these modes are correct.  In
Declaration::tcheck() it is used to find the primary function template
for a specialization that has just been declared so that the forwarded
specialization may be registered into the instantiation list of the
primary.  In TD_func::itcheck() it is used twice, both times to find
the primary for a template function that was just declared.  It no
longer asserts failure if the variable it finds is not a primary
template, as it seems it can sometimes find a function member of a
template class primary, which is not a primary template technically,
but seems to act as one for such functions; I'm not too sure I have
that all correct.

Had to add two arguments to Env::instantiateTemplate():

   1) Variable *bestV

When a template primary or specialization is declared but not defined,
and then instantiations are made of it, and only later is the
definition provided, then when the definition is found, we
re-instantiate the template to re-create these instantiations with
real bodies.  This second time, we want to tell instantiateTemplate()
the bestV to choose from the list of possibilities, as we want it to
choose the template that was forwarded, not just any template that may
happen to fit better from the current list, especially since that
would include an exact match for the previous instantiation (from the
declaration) that we are trying to replace.

   2) bool addTheInstV

This defaults to true, but for the same situation mentioned above, we
do not want the instantiation added to the primary's instantiation
list because the first instantiation (from the declaration) is already
there and this would be an error to have it twice.  We simply point it
at the new instantiation (from the definition).

TemplateInfo objects now have a new field:
  Variable *instantiatedFrom;
This is set by instantiate template so that we know what primary or
specialization an instantiation was instantiated from.  This is handy
as if that primary or specialization was forwarded and later we find
its definition, we can search the list of instantiations for any that
were instantiated from this newly-defined template and re-instantiate
them.

class StackMaintainer now allows a NULL argument which turns it into a
no-op.  Sometimes you can't map the need to call this class completely
onto the control flow, and so some dataflow is involved; since the
dtor for this class is used as a kind of finally statement, we can't
nest its construction in an 'if' statement!  Instead pass in NULL if
you want a no-op effect.

Created a generic MatchTypes::subtractFlags() for subtracting
cv-qualifier flag sets, as this happens more than one place now.

Changed the semantics of MM_WILD and MM_ISO in the presence of
cv-qualifier flags to basically be more correct (as far as I can
tell).  In the case of MM_WILD we don't match if there is a qualifier
on the right that is not present on the left; that is a const type
variable can't match a non-const one.  In the case of MM_ISO, we have
always done the flag subtraction when creating the bindings as in
MM_BIND, but we now check that if there is a binding that the value
bound to matches the flags of the over-determined second binding
exactly.

I was previously intending to mask off non-normal (Scott's) qualifiers
(that is, not CV_CONST or CV_VOLATILE) but sometimes I was using
CV_CONST | CV_NONE instead of CV_CONST | CV_VOLATILE.  Unified them
all into one mask and checked it is right.
  CVFlags const MatchTypes::normalCvFlagMask = CV_CONST | CV_VOLATILE;

Fixed bug where I was not taking cv-qualifiers into account when
checking for the matching of CVAtomicType-s that were instantiations
of class templates, but only for other kinds of CVAtomicType-s; now I
do it for both and in the same place.  There are in fact two function
templates in in/big/nsCLiveconnectFactory.i that differ only because
of this.